+++ src/NuGetGallery/App_Start/AppActivator.cs
@@ -186,8 +186,8 @@ namespace NuGetGallery
 
             var newScriptBundle = new ScriptBundle("~/Scripts/gallery/site.min.js")
                 .Include("~/Scripts/gallery/jquery-1.12.4.js")
-                .Include("~/Scripts/gallery/jquery.validate.js")
-                .Include("~/Scripts/gallery/jquery.validate.unobtrusive.js")
+                .Include("~/Scripts/gallery/jquery.validate-1.16.0.js")
+                .Include("~/Scripts/gallery/jquery.validate.unobtrusive-3.2.6.js")
                 .Include("~/Scripts/gallery/knockout-3.4.2.js")
                 .Include("~/Scripts/gallery/bootstrap.js")
                 .Include("~/Scripts/gallery/moment-2.18.1.js")
